Add functionality to show one time paste
