+++ client/src/components/WarningButton.js
@@ -1,0 +1,11 @@
+import styled from '@emotion/styled';
+import Button from './Button';
+
+const WarningButton = styled(Button)`
+  background: red;
+  color: whitesmoke;
+  margin: 20px 0;
+  padding: 10px 20px;
+`;
+
+export default WarningButton;
