+++ client/src/pages/Paste.js
@@ -1,22 +1,16 @@
 import React from 'react';
 import Loading from '../components/Loading';
 import Alert from '../components/Alert';
-import styled from '@emotion/styled';
 import useGetPaste from '../hooks/useGetPaste';
 import Button from '../components/Button';
 import FullContainer from '../components/FullContainer';
 import PropTypes from 'prop-types';
 import PasteBody from '../components/PasteBody';
-
-const WarningButton = styled(Button)`
-  background: red;
-  color: whitesmoke;
-  margin: 20px 0;
-  padding: 10px 20px;
-`;
+import WarningButton from '../components/WarningButton';
 
 function Paste({ match }) {
   const [{ paste, error, loading }, doGet] = useGetPaste(match.params.pasteId);
+  const [oneTimeActive, setOneTimeActive] = React.useState(false);
 
   return (
     <FullContainer>
@@ -32,8 +26,15 @@ function Paste({ match }) {
       {paste && !paste.oneTimeView && <PasteBody paste={paste} />}
       {paste && paste.oneTimeView && (
         <>
-          <label>You can see it only once! Are you ready?</label>
-          <WarningButton>YES!!</WarningButton>
+          {!oneTimeActive && (
+            <>
+              <label>You can see it only once. Are you ready?</label>
+              <WarningButton onClick={() => setOneTimeActive(!oneTimeActive)}>
+                YES!!!
+              </WarningButton>
+            </>
+          )}
+          {oneTimeActive && <PasteBody paste={paste} />}
         </>
       )}
     </FullContainer>
